refactor "Switch the Lamp" task
