+++ Standard Library Types/Smart Pointers/Switch the Lamp/Cargo.toml
