+++ Standard Library Types/Smart Pointers/Switch the Lamp/src/lib.rs
