+++ Standard Library Types/Smart Pointers/lesson-info.yaml
@@ -1,4 +1,4 @@
 content:
 - Intro
 - Box Task
-- Use Rc and RefCell
+- Switch the Lamp
